Refactor vision tower loading in multimodal encoders

+++ llava/model/multimodal_encoder/open_clip_encoder.py
@@ -1,10 +1,9 @@
 import torch
 import torch.nn as nn
 import torchvision
-import torch.utils.checkpoint as checkpoint
 from transformers import CLIPImageProcessor
 from llava.utils import rank0_print
-from llava.utils import rank0_print
+
 try:
     import open_clip
     from open_clip.transformer import _expand_token
@@ -40,7 +39,7 @@ class OpenCLIPVisionTower(nn.Module):
     def load_model(self, device_map="auto"):
         rank0_print(f"Loading OpenCLIP model: {self.model_name}")
         rank0_print(f"Pretrained: {self.pretrained}")
-        vision_tower, _, image_processor = open_clip.create_model_and_transforms(model_name=self.model_name, pretrained=self.pretrained, precision="bf16", device="cuda")
+        vision_tower, _, image_processor = open_clip.create_model_and_transforms(model_name=self.model_name, pretrained=self.pretrained, precision="fp32", device="cuda")
 
         resize_transform = [t for t in image_processor.transforms if isinstance(t, torchvision.transforms.Resize)][0]
         normalize_transform = [t for t in image_processor.transforms if isinstance(t, torchvision.transforms.Normalize)][0]
